Refactor operations out of handler
